+++ internal/app/zscaler-terraformer/cmd/generate.go
@@ -43,6 +43,7 @@ import (
 	"github.com/zscaler/zscaler-sdk-go/v2/zia/services/user_authentication_settings"
 	"github.com/zscaler/zscaler-sdk-go/v2/zpa/services/appconnectorgroup"
 	"github.com/zscaler/zscaler-sdk-go/v2/zpa/services/applicationsegment"
+	"github.com/zscaler/zscaler-sdk-go/v2/zpa/services/applicationsegmentinspection"
 	"github.com/zscaler/zscaler-sdk-go/v2/zpa/services/applicationsegmentpra"
 	"github.com/zscaler/zscaler-sdk-go/v2/zpa/services/appservercontroller"
 	"github.com/zscaler/zscaler-sdk-go/v2/zpa/services/bacertificate"
@@ -71,6 +72,7 @@ var allGeneratableResources = []string{
 	"zpa_application_server",
 	"zpa_application_segment",
 	"zpa_application_segment_browser_access",
+	"zpa_application_segment_inspection",
 	"zpa_application_segment_pra",
 	"zpa_ba_certificate",
 	"zpa_segment_group",
@@ -412,6 +414,15 @@ func generate(cmd *cobra.Command, writer io.Writer, resourceType string) {
 	case "zpa_application_segment_browser_access":
 		zpaClient := api.zpa.browseraccess
 		jsonPayload, _, err := browseraccess.GetAll(zpaClient)
+		if err != nil {
+			log.Fatal(err)
+		}
+		m, _ := json.Marshal(jsonPayload)
+		resourceCount = len(jsonPayload)
+		_ = json.Unmarshal(m, &jsonStructData)
+	case "zpa_application_segment_inspection":
+		zpaClient := api.zpa.applicationsegmentinspection
+		jsonPayload, _, err := applicationsegmentinspection.GetAll(zpaClient)
 		if err != nil {
 			log.Fatal(err)
 		}