+++ CHANGELOG.md
@@ -18,8 +18,9 @@
 - [PR #206](https://github.com/zscaler/zscaler-terraformer/pull/206). Fixed issues with credential input support via credential file and inline credential authentication.
 
 ### Deprecations
-- [PR #206](https://github.com/zscaler/zscaler-terraformer/pull/206) Deprecated the following ZIA resource:
+- [PR #206](https://github.com/zscaler/zscaler-terraformer/pull/206) Deprecated the following ZIA resources:
     - `zia_user_management`
+    - `zia_admin_users`
 
 ### Internal Changes
 - [PR #206](https://github.com/zscaler/zscaler-terraformer/pull/206). The tool introduced two new internal environment variables for development purposes: ``ZPA_PROVIDER_NAMESPACE`` and ``ZIA_PROVIDER_NAMESPACE``. By setting these enviornment variables it allows a developer to leverage a local Terraform Provider installation for testing purposes. i.e ``export ZPA_PROVIDER_NAMESPACE=zscaler.com/zpa/zpa``. This will force the tool to use a local provider binary installation. If not set, then the tool will download the latest version of the provider from the Terraform registry.